🐛 fix: Leaderboard 탭명 BE 따라가기
