+++ app/src/pages/LeaderBoardPage/tabs/index.ts
@@ -1,4 +1,4 @@
-export * from './CoalitionScoreRankTab';
-export * from './EvalCountRankTab';
-export * from './ExpIncrementRankTab';
-export * from './LevelRankTab';
+export * from './LeaderboardCoalitionScoreTab';
+export * from './LeaderboardEvalCountTab';
+export * from './LeaderboardExpIncrementTab';
+export * from './LeaderboardLevelTab';
